services/auth/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pyjwt==2.8.0
bcrypt==4.1.2
structlog==24.1.0
